Remove call to GetStatupInfoA
Special handling of last argument to ShowWindow (Called by
CreateWindowEx) is not necessary. Therefore drop call and save about 80
bytes in the normal link and about 33 bytes using crinkler.

Tested using a shortcut and setting the run option to start with the
window either maximized or minimized.

Also tested using the test program provided by Davepl in the discussion

Signed-off-by: Charles Stevens <chazste@yahoo.com>